# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Marketing Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Brand DNA ─────────────────────────────────────────────────────────────
...
  /clients/{clientId}/brand-assets:
    get:
      operationId: listBrandAssets
      tags: [brand-assets]
      summary: List brand assets for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Brand assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandAsset"
  /clients/{clientId}/brand-assets/{assetId}:
    delete:
      operationId: deleteBrandAsset
      tags: [brand-assets]
      summary: Delete a brand asset
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Uploads ───────────────────────────────────────────────────────────────
...
  /clients/{clientId}/posts/{postId}/images:
    get:
      operationId: listPostImages
      tags: [images]
      summary: List images for a post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Image"
    post:
      operationId: saveImage
      tags: [images]
      summary: Save an image record
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveImageBody"
      responses:
        "201":
          description: Saved image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"

  # ── Content Memory ─────────────────────────────────────────────────────────
...
  /clients/{clientId}/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /clients/{clientId}/campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign with its posts
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign with posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Social Accounts ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        avatarInitials:
          type: string
        logoUrl:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, color, avatarInitials, createdAt, updatedAt]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required: [name]
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        logoUrl:
          type: string

    # ── Brand DNA ────────────────────────────────────────────────────────────
...
    BrandAsset:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        assetType:
          type: string
          enum: [logo, color_palette, reference_image, brochure, sample_post]
        fileUrl:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required: [id, clientId, assetType, fileUrl, createdAt]

    # ── Upload ───────────────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        campaignId:
          type: string
        topic:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        selectedImageUrl:
          type: string
        status:
          type: string
          enum: [draft, approved, scheduled, published, failed, rejected]
        scheduledAt:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog, newsletter]
        postType:
          type: string
          enum: [social, blog, newsletter]
        title:
          type: string
        longFormBody:
          type: string
        generationStatus:
          type: string
          enum: [generating, ready, failed]
        imagePrompt:
          type: string
        publishedAt:
          type: string
        publishedUrl:
          type: string
        publishError:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, topic, caption, status, postType, createdAt, updatedAt]
...
    Image:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        postId:
          type: string
        url:
          type: string
        provider:
          type: string
          enum: [openai, google]
        status:
          type: string
          enum: [selected, rejected, pending]
        prompt:
          type: string
        createdAt:
          type: string
      required: [id, clientId, postId, url, provider, status, createdAt]
    SaveImageBody:
      type: object
      properties:
        url:
          type: string
        provider:
          type: string
          enum: [openai, google]
        status:
          type: string
          enum: [selected, rejected, pending]
        prompt:
          type: string
      required: [url, provider, status]

    # ── Content Memory ────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        name:
          type: string
        goal:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        platforms:
          type: string
        status:
          type: string
          enum: [draft, active, completed, paused]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, name, status, createdAt, updatedAt]
    CampaignDetail:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        name:
          type: string
        goal:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        platforms:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required: [id, clientId, name, status, createdAt, updatedAt, posts]
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        platforms:
          type: string
        status:
          type: string
          enum: [draft, active, completed, paused]
      required: [name]
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        platforms:
          type: string
        status:
          type: string
          enum: [draft, active, completed, paused]

    # ── AI Suggestions ────────────────────────────────────────────────────────
...